// Frontend/client/model/deliveryNote.ts
/**
 * BillingSoftware.Web
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Company } from './company';
import { DocumentInformations } from './documentInformations';
import { Status } from './status';

export interface DeliveryNote { 
    deliveryNoteNumber?: string;
    deliveryNoteDate: Date;
    status?: Status;
    subject?: string;
    headerText?: string;
    flowText?: string;
    companyId?: string;
    documentInformationsId?: string;
    documentInformations?: DocumentInformations;
    company?: Company;
    id?: string;
    rowVersion?: string;
}